Spin off QClass
